* moved AudioService to a separate file * code cleanup
